moved example files to ./*-example directories
