+++ frontend/app-example/(tabs)/explore.tsx
@@ -1,11 +1,11 @@
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { StyleSheet, Image, Platform } from 'react-native';
 
-import { Collapsible } from '@/components/Collapsible';
-import { ExternalLink } from '@/components/ExternalLink';
-import ParallaxScrollView from '@/components/ParallaxScrollView';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
+import { Collapsible } from '@/components-example/Collapsible';
+import { ExternalLink } from '@/components-example/ExternalLink';
+import ParallaxScrollView from '@/components-example/ParallaxScrollView';
+import { ThemedText } from '@/components-example/ThemedText';
+import { ThemedView } from '@/components-example/ThemedView';
 
 export default function TabTwoScreen() {
   return (
