+++ frontend/app-example/(tabs)/index.tsx
@@ -1,9 +1,9 @@
 import { Image, StyleSheet, Platform } from 'react-native';
 
-import { HelloWave } from '@/components/HelloWave';
-import ParallaxScrollView from '@/components/ParallaxScrollView';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
+import { HelloWave } from '@/components-example/HelloWave';
+import ParallaxScrollView from '@/components-example/ParallaxScrollView';
+import { ThemedText } from '@/components-example/ThemedText';
+import { ThemedView } from '@/components-example/ThemedView';
 
 export default function HomeScreen() {
   return (
