+++ frontend/components-example/ParallaxScrollView.tsx
@@ -7,7 +7,7 @@ import Animated, {
   useScrollViewOffset,
 } from 'react-native-reanimated';
 
-import { ThemedView } from '@/components/ThemedView';
+import { ThemedView } from '@/components-example/ThemedView';
 
 const HEADER_HEIGHT = 250;
