+++ frontend/components-example/navigation/TabBarIcon.tsx
